starting to work on checklist
for now i will keep the progress in this branch. i'll create another one
for the rework of the application in different activites and fragments
and when all works fine i'll merge everything in the master.

added new layout for the checklist sub dialog

crud operation should work even with checklist

i have some doubts on the methods that convert a list to a string and
viceversa but they are a decent placeholder for now

method translate() of class Importance is essentially an override of
Object.toString() so i made it as that

the apk file introduced is a previous working (kinda) version of the app
(it should be updated at the precedent commit)